use layout in route for edit
